♻️ refactor: App.tsx 나누기(대분류), ...ing
